passage de l'ensemble des API en Long plutôt que Integer pour la clé des entités
git-svn-id: svn+ssh://svn.projects.openwide.fr/svnroot/owsi-tools/trunk/owsi-core@1057 d3474844-eb8e-4abc-b058-2b321fed648b

+++ owsi-core-components/owsi-core-component-jpa-more/src/main/java/fr/openwide/core/jpa/more/business/audit/dao/AbstractAuditDaoImpl.java
@@ -37,7 +37,7 @@ import fr.openwide.core.jpa.more.business.audit.model.AbstractAudit_;
  * 
  * @author Open Wide
  */
-public abstract class AbstractAuditDaoImpl<T extends AbstractAudit> extends GenericEntityDaoImpl<Integer, T> 
+public abstract class AbstractAuditDaoImpl<T extends AbstractAudit> extends GenericEntityDaoImpl<Long, T> 
 	implements IAuditDao<T> {
 
 	/**
@@ -47,7 +47,7 @@ public abstract class AbstractAuditDaoImpl<T extends AbstractAudit> extends Gene
 	}
 
 	@Override
-	public GenericEntity<?, ?> getGenericEntity(Class<? extends GenericEntity<?, ?>> clazz, Integer id) {
+	public GenericEntity<?, ?> getGenericEntity(Class<? extends GenericEntity<?, ?>> clazz, Long id) {
 		return (GenericEntity<?, ?>) getEntity(clazz, id);
 	}
