+++ owsi-core-components/owsi-core-component-jpa-more/src/main/java/fr/openwide/core/jpa/more/business/parameter/dao/ParameterDaoImpl.java
@@ -6,7 +6,7 @@ import fr.openwide.core.jpa.business.generic.dao.GenericEntityDaoImpl;
 import fr.openwide.core.jpa.more.business.parameter.model.Parameter;
 
 @Repository("parameterDao")
-public class ParameterDaoImpl extends GenericEntityDaoImpl<Integer, Parameter>
+public class ParameterDaoImpl extends GenericEntityDaoImpl<Long, Parameter>
 		implements IParameterDao {
 
 }
