+++ owsi-core-components/owsi-core-component-jpa-more/src/test/java/fr/openwide/core/test/jpa/more/business/entity/dao/TestEntityDaoImpl.java
@@ -6,6 +6,6 @@ import fr.openwide.core.jpa.business.generic.dao.GenericEntityDaoImpl;
 import fr.openwide.core.test.jpa.more.business.entity.model.TestEntity;
 
 @Repository("testEntityDao")
-public class TestEntityDaoImpl extends GenericEntityDaoImpl<Integer, TestEntity> implements ITestEntityDao {
+public class TestEntityDaoImpl extends GenericEntityDaoImpl<Long, TestEntity> implements ITestEntityDao {
 
 }
