+++ owsi-core-examples/owsi-core-example-jpa/src/main/java/fr/openwide/jpa/example/business/person/dao/PersonDaoImpl.java
@@ -6,7 +6,7 @@ import fr.openwide.core.jpa.business.generic.dao.GenericEntityDaoImpl;
 import fr.openwide.jpa.example.business.person.model.Person;
 
 @Repository("personDao")
-public class PersonDaoImpl extends GenericEntityDaoImpl<Integer, Person> implements PersonDao {
+public class PersonDaoImpl extends GenericEntityDaoImpl<Long, Person> implements PersonDao {
 
 	public PersonDaoImpl() {
 		super();
